+++ src/Config/Util/RuntimeType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Shlinkio\Shlink\Installer\Config\Util;
+
+enum RuntimeType: string
+{
+    case ASYNC = 'async';
+    case REGULAR = 'regular';
+}