+++ src/Model/ShlinkInitConfig.php
@@ -14,9 +14,8 @@ final class ShlinkInitConfig
          * False: Do not generate an initial API key.
          * String: Use provided value as the initial API key.
          * Null: Auto-generate a random initial API key.
-         * @todo Change to string|false|null once PHP 8.1 is no longer supported
          */
-        public readonly string|bool|null $generateApiKey,
+        public readonly false|string|null $generateApiKey,
         public readonly bool $downloadGeoLiteDb,
     ) {
     }